Pull out Eval pretty printing
